Move some classes; add some docs
