+++ lib/graphql/schema_member.rb
@@ -1,6 +1,12 @@
 # frozen_string_literal: true
 module GraphQL
+  # The base class for things that make up the schema,
+  # eg objects, enums, scalars.
+  #
+  # @api private
   class SchemaMember
+    # Adds a layer of caching over user-supplied `.to_graphql` methods.
+    # Users override `.to_graphql`, but all runtime code should use `.graphql_definition`.
     module CachedGraphQLDefinition
       # A cached result of {.to_graphql}.
       # It's cached here so that user-overridden {.to_graphql} implementations
@@ -10,46 +16,11 @@ module GraphQL
       end
     end
 
-    # Shared code for Object and Interface
-    module HasFields
-      # Define a field on this object
-      def field(*args, &block)
-        field_defn = field_class.new(*args, &block)
-        add_field(field_defn)
-        nil
-      end
-
-      def fields
-        all_fields = own_fields
-        inherited_fields = (superclass.is_a?(HasFields) ? superclass.fields : [])
-        # Remove any inherited fields which were overridden on this class:
-        inherited_fields.each do |inherited_f|
-          if all_fields.none? {|f| f.name == inherited_f.name }
-            all_fields << inherited_f
-          end
-        end
-        all_fields
-      end
-
-      # Register this field with the class, overriding a previous one if needed
-      def add_field(field_defn)
-        fields.reject! { |f| f.name == field_defn.name }
-        own_fields << field_defn
-      end
-
-      def own_fields
-        @own_fields ||= []
-      end
-
-      def field_class
-        self::Field
-      end
-    end
-
     class << self
       include CachedGraphQLDefinition
 
-      # Make the class act like its corresponding type (eg `connection_type`)
+      # Delegate to the derived type definition if possible.
+      # This is tricky because missing methods cause the definition to be built & cached.
       def method_missing(method_name, *args, &block)
         if graphql_definition.respond_to?(method_name)
           graphql_definition.public_send(method_name, *args, &block)
@@ -58,10 +29,15 @@ module GraphQL
         end
       end
 
+      # Check if the derived type definition responds to the method
+      # @return [Boolean]
       def respond_to_missing?(method_name, incl_private = false)
         graphql_definition.respond_to?(method_name, incl_private) || super
       end
 
+      # Call this with a new name to override the default name for this schema member; OR
+      # call it without an argument to get the name of this schema member
+      # @param new_name [String]
       # @return [String]
       def graphql_name(new_name = nil)
         if new_name
@@ -71,6 +47,9 @@ module GraphQL
         end
       end
 
+      # Call this method to provide a new description; OR
+      # call it without an argument to get the description
+      # @param new_description [String]
       # @return [String]
       def description(new_description = nil)
         if new_description
@@ -92,29 +71,9 @@ module GraphQL
         NonNullTypeProxy.new(self)
       end
     end
-
-    class ListTypeProxy
-      include GraphQL::SchemaMember::CachedGraphQLDefinition
-
-      def initialize(member)
-        @member = member
-      end
-
-      def to_graphql
-        @member.graphql_definition.to_list_type
-      end
-    end
-
-    class NonNullTypeProxy
-      include GraphQL::SchemaMember::CachedGraphQLDefinition
-
-      def initialize(member)
-        @member = member
-      end
-
-      def to_graphql
-        @member.graphql_definition.to_non_null_type
-      end
-    end
   end
 end
+
+require 'graphql/schema_member/list_type_proxy'
+require 'graphql/schema_member/non_null_type_proxy'
+require 'graphql/schema_member/has_fields'
